+++ src/components/ActionBar.jsx
@@ -1,44 +1,16 @@
 import "./ActionBar.scss";
 import { useEffect, useState } from "react";
 
-import {getReactions, getRecipient} from "../api/apis.js";
-import arrowBtn from "../assets/icons/btn_arrow.svg";
-import emojiAdd from "../assets/icons/emoji-add.png";
-import shareBtn from "../assets/icons/Union.svg";
+import { getRecipient } from "../api/apis.js";
 
-import Button from "./Button";
-import Emoji from "./Emoji";
+import ActionBarBtnsGroup from "./ActionBarBtnsGroup.jsx"
 import Profile from "./Profile";
-import ReactionWindow from "./ReactionWindow";
-
-const INITIAL_LIMIT = 11;
-const LOAD_MORE_LIMIT = 8;
 
 const ActionBar = ({ recipientId, messagesData }) => {
   const [recipientData, setRecipientData] = useState({});
 
-  const [reactionsData, setReactionsData] = useState({ results: [] });
-
-  const [isOpened, setIsOpened] = useState(false);
-  const [offset, setOffset] = useState(0);
-
   const images =
     messagesData?.results?.map((message) => message.profileImageURL) ?? [];
-
-  const getInitReactions = async () => {
-    try {
-      const reactionsResData = await getReactions({
-        recipientId,
-        limit: INITIAL_LIMIT,
-        offset: 0,
-      });
-      setReactionsData(reactionsResData);
-      setOffset(INITIAL_LIMIT);
-    } catch (e) {
-      console.log(e.message);
-      alert("오류가 발생했습니다.");
-    }
-  };
 
   const getRecipientData = async () => {
     try {
@@ -52,33 +24,7 @@ const ActionBar = ({ recipientId, messagesData }) => {
 
   useEffect(() => {
     getRecipientData();
-    getInitReactions();
   }, [recipientId]);
-
-  const onClickGetReactions = () => {
-    setIsOpened(prev => !prev);
-    if (Number(reactionsData?.results.length) > 11) {
-      getInitReactions();
-    }
-  };
-
-  const onClickLoadMore = async () => {
-    try {
-      const loadMoreRes = await getReactions({
-        recipientId,
-        limit: LOAD_MORE_LIMIT,
-        offset,
-      });
-      setOffset((prev) => prev + LOAD_MORE_LIMIT);
-      setReactionsData({
-        ...reactionsData,
-        results: [...reactionsData.results, ...loadMoreRes.results],
-      });
-    } catch (e) {
-      console.log(e.message);
-      alert("오류가 발생했습니다.");
-    }
-  };
 
   return (
     <div className="ActionBar">
@@ -93,35 +39,10 @@ const ActionBar = ({ recipientId, messagesData }) => {
             </div>
           </div>
           <div className="divider"></div>
-          <div className="reaction-wrapper">
-            <div className="badges">
-              {recipientData?.topReactions?.map((emoji) => (
-                <Emoji key={emoji.id} emoji={emoji.emoji} count={emoji.count} />
-              ))}
-            </div>
-            <button className="reaction-window-btn" onClick={onClickGetReactions}>
-              <img
-                src={arrowBtn}
-                style={{
-                  transform: isOpened ? "rotate(0deg)" : "rotate(180deg)",
-                }}
-              />
-            </button>
-            {isOpened && (
-              <ReactionWindow
-                data={reactionsData}
-                onClickLoadMore={onClickLoadMore}
-                onClickGetReactions={onClickGetReactions}
-              />
-            )}
-          </div>
-          <div className="btn-wrapper">
-            <Button icon={emojiAdd} size="sm" variant="outlined">
-              추가
-            </Button>
-            <div className="divider"></div>
-            <Button icon={shareBtn} size="sm" variant="outlined" />
-          </div>
+          <ActionBarBtnsGroup 
+            recipientId={recipientId} 
+            recipientData={recipientData}
+          />
         </div>
       </div>
     </div>
